requirements.txt
flask
gunicorn
cs50
temp
zipfile36
authlib
google-api-python-client
google-auth-oauthlib
psycopg2-binary
